requirements.txt
appnope==0.1.2
APScheduler==3.6.3
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
distro==1.6.0
executing==0.8.2
html5lib==1.1
idna==3.3
ipython==8.0.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.1
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
Pygments==2.11.2
pyTelegramBotAPI==4.3.1
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
stack-data==0.1.3
tabula-py==2.3.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
